+++ angular/lib/src/core/di.dart
@@ -1,4 +1,5 @@
 /// A dependency Injection container.
+export "../di/errors.dart" show InjectionError, NoProviderError;
 export "../di/injector/injector.dart" show Injector;
 export "../di/injector/runtime.dart" show ReflectiveInjector;
 export "../di/module.dart" hide internalModuleToList;