+++ angular_router/test/on_navigation_start_test.dart
@@ -79,7 +79,11 @@ Stream navigate(Router router, String path) => StreamGroup.merge([
 const canDeactivateToken = const OpaqueToken<bool>('canDeactivateToken');
 const canNavigateToken = const OpaqueToken<bool>('canNavigateToken');
 
-@Component(selector: 'home', template: '')
+@Component(
+  selector: 'home', template: '',
+  // TODO(b/71710685): Change to `Visibility.local` to reduce code size.
+  visibility: Visibility.all,
+)
 class HomeComponent implements CanDeactivate, CanNavigate {
   final bool _canDeactivate;
   final bool _canNavigate;
@@ -98,7 +102,11 @@ class HomeComponent implements CanDeactivate, CanNavigate {
   Future<bool> canNavigate() => new Future.value(_canNavigate);
 }
 
-@Component(selector: 'destination', template: '')
+@Component(
+  selector: 'destination', template: '',
+  // TODO(b/71710685): Change to `Visibility.local` to reduce code size.
+  visibility: Visibility.all,
+)
 class DestinationComponent {}
 
 @Component(
@@ -106,6 +114,8 @@ class DestinationComponent {}
   template: '<router-outlet [routes]="routes"></router-outlet>',
   directives: const [RouterOutlet],
   providers: const [routerProvidersTest],
+  // TODO(b/71710685): Change to `Visibility.local` to reduce code size.
+  visibility: Visibility.all,
 )
 class TestComponent {
   final Router router;